@@ -3,45 +3,40 @@
     "version": "3.0.0",
     "scripts": {
         "ng": "ng",
-        "start": "ng serve --host=127.0.0.1 --ssl --ssl-cert ssl\\localhost.pem --ssl-key ssl\\localhost-key.pem",
+        "start": "ng serve --hmr --host=127.0.0.1 --ssl --ssl-cert ssl\\localhost.pem --ssl-key ssl\\localhost-key.pem",
         "build": "ng build",
         "lint": "eslint ./"
     },
     "private": true,
     "dependencies": {
-        "@angular/animations": "15.2.9",
-        "@angular/common": "15.2.9",
-        "@angular/compiler": "15.2.9",
-        "@angular/core": "15.2.9",
-        "@angular/forms": "15.2.9",
-        "@angular/material": "14.2.7",
-        "@angular/material-moment-adapter": "14.2.7",
-        "@angular/platform-browser": "15.2.9",
-        "@angular/platform-browser-dynamic": "15.2.9",
-        "@angular/router": "15.2.9",
-        "@zxing/browser": "0.1.3",
-        "@zxing/library": "0.20.0",
-        "@zxing/ngx-scanner": "3.9.0",
+        "@angular/animations": "16.2.9",
+        "@angular/common": "16.2.9",
+        "@angular/compiler": "16.2.9",
+        "@angular/core": "16.2.9",
+        "@angular/forms": "16.2.9",
+        "@angular/material": "16.2.8",
+        "@angular/material-moment-adapter": "16.2.8",
+        "@angular/platform-browser": "16.2.9",
+        "@angular/platform-browser-dynamic": "16.2.9",
+        "@angular/router": "16.2.9",
+        "crypto-js": "4.2.0",
         "dexie": "3.2.4",
         "guid-typescript": "1.0.9",
-        "jwt-decode": "3.1.2",
-        "pdfmake": "0.2.7",
-        "pretty-print-json": "2.0.1",
-        "primeng": "15.4.1",
+        "primeng": "16.5.0",
         "rxjs": "7.8.1",
-        "stylelint": "15.6.1",
-        "tslib": "2.5.0",
+        "stylelint": "15.10.3",
+        "tslib": "2.6.2",
         "zone.js": "0.13.0"
     },
     "devDependencies": {
-        "@angular-devkit/build-angular": "15.2.8",
-        "@angular/cli": "15.2.8",
-        "@angular/compiler-cli": "15.2.9",
-        "@types/pdfmake": "0.2.2",
-        "@typescript-eslint/eslint-plugin": "5.59.2",
-        "@typescript-eslint/parser": "5.59.2",
-        "eslint": "8.39.0",
-        "stylelint-config-standard": "33.0.0",
-        "typescript": "4.9.5"
+        "@angular-devkit/build-angular": "16.2.6",
+        "@angular/cli": "16.2.6",
+        "@angular/compiler-cli": "16.2.9",
+        "@types/pdfmake": "0.2.5",
+        "@typescript-eslint/eslint-plugin": "6.7.5",
+        "@typescript-eslint/parser": "6.7.5",
+        "eslint": "8.51.0",
+        "stylelint-config-standard": "34.0.0",
+        "typescript": "5.1.6"
     }
 }
