@@ -11,6 +11,6 @@ export interface CheckInPassengerWriteDto {
     birthdate: string
     remarks: string
     specialCare: string
-    isCheckedIn: boolean
+    isBoarded: boolean
 
 }
