@@ -3,7 +3,7 @@ import { Injectable } from '@angular/core'
 import { Observable } from 'rxjs'
 // Custom
 import { HttpDataService } from 'src/app/shared/services/http-data.service'
-import { NationalityDropdownVM } from '../view-models/nationality-dropdown-vm'
+import { NationalityDropdownVM } from '../view-models/nationality-autocomplete-vm'
 import { environment } from 'src/environments/environment'
 
 @Injectable({ providedIn: 'root' })
@@ -16,8 +16,8 @@ export class NationalityService extends HttpDataService {
 
     //#region public methods
 
-    getActive(): Observable<NationalityDropdownVM[]> {
-        return this.http.get<NationalityDropdownVM[]>(environment.apiUrl + '/nationalities/getActive')
+    getAutoComplete(): Observable<NationalityDropdownVM[]> {
+        return this.http.get<NationalityDropdownVM[]>(environment.apiUrl + '/nationalities/getAutoComplete')
     }
 
     //#endregion
