@@ -1,8 +1,6 @@
-import { ChangeDetectorRef, Component, HostListener } from '@angular/core'
+import { ChangeDetectorRef, Component } from '@angular/core'
 import { NavigationCancel, NavigationEnd, NavigationError, NavigationStart, Router } from '@angular/router'
 // Custom
-import { AccountService } from '../shared/services/account.service'
-import { ConnectedUser } from '../shared/classes/connected-user'
 import { LoadingSpinnerService } from '../shared/services/loading-spinner.service'
 import { environment } from 'src/environments/environment'
 import { routeAnimation } from '../shared/animations/animations'
@@ -22,7 +20,7 @@ export class AppComponent {
 
     //#endregion
 
-    constructor(private accountService: AccountService, private changeDetector: ChangeDetectorRef, private router: Router, private loadingSpinnerService: LoadingSpinnerService) {
+    constructor(private changeDetector: ChangeDetectorRef, private loadingSpinnerService: LoadingSpinnerService, private router: Router) {
         this.router.events.subscribe((routerEvent) => {
             if (routerEvent instanceof NavigationStart) {
                 this.isLoading = true
@@ -33,21 +31,12 @@ export class AppComponent {
         })
     }
 
-    //#region listeners
-
-    @HostListener('window:beforeunload', ['$event']) beforeUnloadHander(): any {
-        this.accountService.logout()
-    }
-
-    //#endregion
-
     //#region lifecycle hooks
 
     ngOnInit(): void {
         this.initLoadingSpinner()
-        this.setTheme()
-        this.openBroadcastChannel()
-        this.isUserConnected()
+        this.setUserSelect()
+        this.setBackgroundImage()
     }
 
     //#endregion
@@ -59,20 +48,13 @@ export class AppComponent {
             this.isLoading = status == 'start'
             this.changeDetector.detectChanges()
         })
-
     }
 
-    private isUserConnected(): void {
-        if (ConnectedUser.id == undefined && window.location.href.includes('resetPassword') == false) {
-            this.accountService.logout()
-        }
+    private setBackgroundImage(): void {
+        document.getElementById('wrapper').style.backgroundImage = 'url(../../assets/images/themes/background.svg'
     }
 
-    private openBroadcastChannel(): void {
-        new BroadcastChannel('test').postMessage('open')
-    }
-
-    private setTheme(): void {
+    private setUserSelect(): void {
         document.getElementById('main').style.userSelect = environment.cssUserSelect
     }
 
