@@ -4,7 +4,6 @@ import { MatAutocompleteTrigger } from '@angular/material/autocomplete'
 import { MatExpansionPanel } from '@angular/material/expansion'
 import { Observable, Subject, defer, finalize } from 'rxjs'
 import { Router } from '@angular/router'
-import { Table } from 'primeng/table'
 import { Title } from '@angular/platform-browser'
 // Custom
 import { MessageLabelService } from './message-label.service'
@@ -36,20 +35,13 @@ export class HelperService {
 
     //#endregion
 
-    constructor(
-        private dialogService: ModalDialogService,
-        private messageLabelService: MessageLabelService,
-        private router: Router,
-        private sessionStorageService: SessionStorageService,
-        private titleService: Title
-    ) { }
+    constructor(private messageLabelService: MessageLabelService, private modalDialogService: ModalDialogService, private router: Router, private sessionStorageService: SessionStorageService, private titleService: Title) { }
 
     //#region public methods
 
-    public doPostSaveFormTasks(message: string, iconType: string, returnUrl: string, form: any, formReset = true, goBack = true): Promise<any> {
+    public doPostSaveFormTasks(message: string, iconType: string, returnUrl: string, goBack: boolean): Promise<any> {
         const promise = new Promise((resolve) => {
-            this.dialogService.open(message, iconType, ['ok']).subscribe(() => {
-                formReset ? form.reset() : null
+            this.modalDialogService.open(message, iconType, ['ok']).subscribe(() => {
                 goBack ? this.router.navigate([returnUrl]) : null
                 resolve(null)
             })
@@ -78,64 +70,12 @@ export class HelperService {
 
     public focusOnField(): void {
         setTimeout(() => {
-            const input = Array.prototype.slice.apply(document.querySelectorAll('input[data-tabindex]'))[0]
-            if (input != null) {
+            const input = Array.prototype.slice.apply(document.querySelectorAll('input[dataTabIndex]'))[0]
+            if (input != null && this.sessionStorageService.getItem('isFirstFieldFocused') == 'true') {
                 input.focus()
                 input.select()
             }
         }, 500)
-    }
-
-    public enableTableFilters(): void {
-        setTimeout(() => {
-            const checkboxes = document.querySelectorAll('.p-checkbox, .p-checkbox-box') as NodeListOf<HTMLElement>
-            checkboxes.forEach(x => {
-                x.classList.remove('disabled')
-            })
-            const datePickers = document.querySelectorAll('.mat-datepicker-toggle, .mat-datepicker-toggle > .mat-button-base > .mat-button-wrapper') as NodeListOf<HTMLElement>
-            datePickers.forEach(x => {
-                x.classList.remove('disabled')
-            })
-            const dropdown = document.querySelectorAll('.p-inputwrapper') as NodeListOf<HTMLElement>
-            dropdown.forEach(x => {
-                x.classList.remove('disabled')
-            })
-            const textFilters = document.querySelectorAll('.p-inputtext')
-            textFilters.forEach(x => {
-                x.classList.remove('disabled')
-            })
-        }, 500)
-    }
-
-    public disableTableFilters(): void {
-        setTimeout(() => {
-            const checkboxes = document.querySelectorAll('.p-checkbox, .p-checkbox-box') as NodeListOf<HTMLElement>
-            checkboxes.forEach(x => {
-                x.classList.add('disabled')
-            })
-            const datePickers = document.querySelectorAll('.mat-datepicker-toggle, .mat-datepicker-toggle > .mat-button-base > .mat-button-wrapper') as NodeListOf<HTMLElement>
-            datePickers.forEach(x => {
-                x.classList.add('disabled')
-            })
-            const dropdown = document.querySelectorAll('.p-inputwrapper') as NodeListOf<HTMLElement>
-            dropdown.forEach(x => {
-                x.classList.add('disabled')
-            })
-            const textFilters = document.querySelectorAll('.p-inputtext')
-            textFilters.forEach(x => {
-                x.classList.add('disabled')
-            })
-        }, 500)
-    }
-
-    public clearTableTextFilters(table: Table, inputs: string[]): void {
-        table.clear()
-        inputs.forEach(input => {
-            table.filter('', input, 'contains')
-        })
-        document.querySelectorAll<HTMLInputElement>('.p-inputtext, .mat-input-element').forEach(box => {
-            box.value = ''
-        })
     }
 
     public flattenObject(object: any): any {
@@ -166,23 +106,10 @@ export class HelperService {
         })
     }
 
-    public sortNestedArray(array: any, property: any): any {
-        property = property.split('.')
-        const len = property.length
-        array.sort((a: any, b: any) => {
-            let i = 0
-            while (i < len) { a = a[property[i]]; b = b[property[i]]; i++ }
-            if (a < b) {
-                return -1
-            } else if (a > b) {
-                return 1
-            } else {
-                return 0
-            }
-        })
-    }
-
     public deepEqual(object1: any, object2: any): boolean {
+        if (object1 == undefined || object2 == undefined) {
+            return false
+        }
         const keys1 = Object.keys(object1)
         const keys2 = Object.keys(object2)
         if (keys1.length !== keys2.length) {
@@ -193,8 +120,7 @@ export class HelperService {
             const val2 = object2[key]
             const areObjects = this.isObject(val1) && this.isObject(val2)
             if (
-                areObjects && !this.deepEqual(val1, val2) ||
-                !areObjects && val1 !== val2
+                areObjects && !this.deepEqual(val1, val2) || !areObjects && val1 !== val2
             ) {
                 return false
             }
@@ -220,22 +146,6 @@ export class HelperService {
         }, 500)
     }
 
-    public unHighlightAllRows(): void {
-        const x = document.querySelectorAll('.p-highlight')
-        x.forEach(row => {
-            row.classList.remove('p-highlight')
-        })
-    }
-
-    public clearTableCheckboxes(): void {
-        setTimeout(() => {
-            const x = document.querySelectorAll('tr td .p-element .p-checkbox .p-checkbox-box .p-checkbox-icon.pi')
-            x.forEach(row => {
-                row.classList.remove('pi-check')
-            })
-        }, 100)
-    }
-
     public scrollToSavedPosition(virtualElement: any, feature: string): void {
         if (virtualElement != undefined) {
             setTimeout(() => {
@@ -249,26 +159,11 @@ export class HelperService {
     }
 
     public openOrCloseAutocomplete(form: FormGroup<any>, element: any, trigger: MatAutocompleteTrigger): void {
-        form.get(element).patchValue('')
         trigger.panelOpen ? trigger.closePanel() : trigger.openPanel()
-    }
-
-    public goBackFromForm(form: FormGroup<any>): any {
-        form.reset()
-        setTimeout(() => {
-            return true
-        }, 1000)
     }
 
     public setTabTitle(feature: string): void {
         this.titleService.setTitle(environment.appName + ': ' + this.messageLabelService.getDescription(feature, 'header'))
-    }
-
-    public calculateDayCount(): number {
-        const elementWidth = document.getElementById('content').clientWidth
-        console.log(elementWidth)
-        const dayCount = Math.trunc(elementWidth / 123.315)
-        return dayCount
     }
 
     public toggleExpansionPanel(panels: QueryList<MatExpansionPanel> | { open: () => any; close: () => any }[], newState: boolean): void {
@@ -279,13 +174,16 @@ export class HelperService {
         })
     }
 
+    public setSidebarsTopMargin(margin: string): void {
+        const sidebars = document.getElementsByClassName('sidebar') as HTMLCollectionOf<HTMLElement>
+        for (let i = 0; i < sidebars.length; i++) {
+            sidebars[i].style.marginTop = margin + 'rem'
+        }
+    }
+
     //#endregion
 
     //#region private methods
-
-    public getLabel(feature: string, id: string): string {
-        return this.messageLabelService.getDescription(feature, id)
-    }
 
     private isObject(object: any): boolean {
         return object != null && typeof object === 'object'
